Changed misspellings and started Celery implementation Changed all instances of Pluggins/pluggins to Plugins/plugins Began adding celery support. A competition will attempt to start the scoring engine when the start time is reached. There is little to no error/edge case checking.

+++ cssefwebfront/config.py
@@ -1,4 +1,5 @@
 from django.apps import AppConfig
+from os.path import isfile
 import settings
 
 class cssefwebfront_config(AppConfig):
@@ -6,15 +7,9 @@ class cssefwebfront_config(AppConfig):
 	verbose_name = 'CSSEF Web Front'
 
 	def ready(self):
-		# Ensure '__init__.py' file lead to pluggin directory
+		# Ensure '__init__.py' file lead to plugin directory
 		cur_path = settings.BASE_DIR + "/"
 		for i in settings.CONTENT_PLUGGINS_PATH.split("/"):
-			if i != "":
+			if i != "" and not isfile(cur_path + "/__init__.py"):
 				cur_path += i + "/"
 				open(cur_path + "__init__.py", 'w').close()
-
-		# Import the signals handler as was done here: http://chriskief.com/2014/02/28/django-1-7-signals-appconfig/
-		# Not really sure if this is the best way to do it...
-		from cssefwebfront import signals
-		
-
